Fix tags with transparency.
Only change the tag if the alpha_s (source alpha) is > 0. If it is 100%
and the BlendMode is Normal, then we use the object's tag, if not, we
combine with the existing tag (in the backdrop/destination) by OR'ing
the bits together. Note that for some Blend modes, the pixel may only
have contributions from the source or from the destination, but we don't
use that because it could change the tag within an object (e.g, gradient
or image) depending on the color value causing unexpected shifts in the
tag value.