Vector devices - work around TrueType subsetting bug corrupting metrics tables
Bug #697376 "Adobe Distiller can't handle EPS files created with eps2write - limitcheck error"

The TrueType writing code for the vector devices continues to produce
problems. We should rewrite this code but realistically that is still
some way off.

In this case the code was attempting to minimise the size of the VMTX
and HMTX tables by only writing out the metrics for the glyphs we were
actually including in the font (unused glyphs remain as references in the
font, but have no outlines).

This is not valid in TrueType, the metrics tables must contain an entry
for each glyph. Looking at the code it should have been clear that this
was nonsense, but apparently it was missed.

We can save a minimal amount of space by writing out the metrics for all
the glyphs up to the last used one, and then writing just a 0 side
bearing for each remaining glyph, but the space saving is small. So I've
chosen simply to disable the subsetting of metrics and copy the entire
table.