pdfwrite - correctly set the BPC for the image downsampling filter
Bug #697944 "Segmentation fault with -sDEVICE=pdfwrite"

The code to call the downsampling filter resizes the input to the filter
to 8BPC, and then resizes the output from the filter back to whatever
the required depth was.

But the stream state for the downsampling filter was left with the
original BPC. This led to the Mitchell filter calculating two parameters
sizeofPixelIn and sizeofPixelOut incorrectly as 0 (4/8 is 0 in integer
arithmetic). Later the code used sizeofPixelOut being 1 to write 1 byte
output data, but if the value was not 1 it wrote 2 bytes of output.

This causes a buffer overflow in the callers buffer, leading to memory
corruption and an eventual seg fault.

This commit simply sets the BPC in the stream state correctly to reflect
the fact that the intermediate data is always 8 bits.