pdfwrite - Don't assume param list strings are NULL-terminated
Bug #697948 "(Add an ID to memory allocations.) breaks PDF/A output"

The bug description is misleading; in fact the problem is that the code
for reading a flag numeric value from a string parameter was assuming
the parameter was NULL_terminated, when in fact it is not.

Depending on exactly how the memory layout happened to be, the bytes
following the parameter may or may not be NULLs, or ASCII digits. If it
so happened that the trailing bytes were ASCII digits, it was possible
for an incorrect Flag value to be parsed, which could lead to an error
being flagged.

This commit just copies the actual number of bytes in the parameter
string to a temporary stack-based character buffer, NULL terminates the
buffer and reads the Flag value from there.