Bug 697748: Fix misspelled alignment attribute in SSE code.
The SSE thresholding code relies on a buffer being 16 byte aligned.
To achieve this we use a compiler attribute to force alignment
to 16 bytes.

MSVC uses "align(16)", but gcc etc uses "aligned(16)".

We were using "align(16)" for both. Fix this in the macro here.

Thanks to Timo Teras for spotting the problem and suggesting the
fix.