Bug 697970: Solve underflow/overflow problems in new scan converter.
The test file asks us to fill a line with one end at 0x7fffffff,
using 'any part of a pixel'. In order to calculate the index table
the existing code 'expands' this vertical area by 0x7f to allow
for rounding within a pixel. This leads to 0x8000007e, which
is of a different sign.

Correspondingly the vertical ranges touched by the lines appear
different in the indexing and marking phases.

To fix this we rejig the line calculations to avoid this
difference by more carefully doing the expansion.

Thanks to Kamil Frankowicz for doing the fuzz testing that lead
to this bug/fix.