PDF interpreter - apply Optional Content /Usage dictionaries
Bug #696076 "Ghost script 9.16/9.15 print PDF layer even if print state is "Never Print""

We currently only apply the /OFF (or /ON) entries from the Optional
Content Properties (/OCProperties) which is quite a high level and hence
low granularity control.

This commit adds processing of the /Usage dictionary; any ordinary
Optional Content (not OCMD) which is not set to /OFF in the OCProperties
is checked for the existence of a /Usage dictionary. We then check for
either a /Print or /View key in the dictionary depending on the setting
of the -dPrinted command line switch. If the relevant key is present
then we check for a ViewState key and use that to determine whether the
Optional Content should be rendered or not.

Note that if -dPrinted is not set on the command line then the interpreter
will attempt to guess. If the current device has OutputFile set then
the interpreter will assume that it should consider 'Print' keys. If it
is not present in the device, then the interpreter will assume that it
should consider 'View' keys.

This causes the PDF file Bug690364.pdf to render radically differently.
This is because much of the content is set to not print, and a small
amount of extra data is set to render only when printed. Because the
cluster uses an OutputFile, we assume that we should behave as if
printing.