Bug 697679: Ensure clist playback correctly creates clip paths.
Ghostscript clip paths consist of 3 parts:

 1) A list of scan converted rectangles that union together to give
 the clipping region.
 2) (Optionally) a path that represents the clipping region.
 3) A path_list structure (a list of paths that when intersected
 together give the required clipping region).

3 is required for high level devices like pdfwrite (and the opvp
device).

Most of the code is at pains to preserve path_list as gx_cpaths are
manipulated. The exception to this appears to be the clist code.
The rectangle list is sent through the clist, but the path is not.

Accordingly, the code to intersect a gx_cpath with another path
checks to see if the path_list is valid on entry. If not, it
synthesizes one from the rectangle list.

This bug is caused by the fact that we end up intersecting a
clip path 120000 times in a pattern clist playback. Each of those
synthesizes a path from the rectangle list.

The fix here is to make the pattern clist playback create the
path_list each time it is replayed, meaning we have < 100 such
operations.

Thanks to Ken for his work on this, and for coming up with the
initial version of this patch.