Bug 697983: avoid marking zero length glyph contours
The file contains many glyphs with stray zero length contours dotted around the
'main body' of the glyph outline. When a text rendering mode involving stroke
was used, these were resulting in the linecaps being drawn, and showing as
extra dots around the gylphs.

Force linecap to 'butt' when using a stroking text rendering mode to avoid this.