pdfwrite - keep using /Subsample for /Indexed images when downsampling
Bug #698000 "Inappropriate choice of downsampling filter for Indexed images"

For Bug #693917 we checked the downsampling factor for the Subsample and
Average downsampling filters and, if it was not an integer or nearly so,
we forced the Bicubic filter instead.

Unfortunately, this doesn't work well for image data in a /Indexed colour
space.

Instead this commit checks the colour space and, if its a /Indexed space,
forces the downsampling factor to the nearest integer, and preserves the
Subsample or Average filter.