PDF interpreter - fix a problem with AcroForm interpretation
No bug for this, noticed in passing while working on a customer bug.

If we turn on AcroForm processing, and we have an AcroFOrm where a /Tx
field has a /V entry, then we would not write the value to the string we
are accumulating. We would only write the value if we had no /V but a
/DV.

Instead the string was left on the stack, leading to an error. This
appears to be due to a logical error when writing the V/DV handling
where the write== was placed inside the if clause, when it should be
outside the clause.