chunk allocator: on free, use original size for single object chunk
On allocation, we decide, based on the requested size (+ house keeping data)
if an object should be in a chunk on its own. But the size recorded can be
larger than that if the chunk was recovered from the free list (because the
free list can contain merged free chunks). In such a case, we also record the
padding bytes.

On freeing, we have to use the object size minus the object padding to
identify single object chunks - to mirror the decision making in the allocation
code.

Fixes the cluster segfault with C410.bin and bitrgbtags