Fix mistake introduced with the pcl glyph data bounds checking code
When I added the bounds checking for PCL glyph data, I changed slightly the
way offsets were calculated for the PXL 'error handler' bitmap font so
we could know the data length.

I messed up, and forgot to fully take account of the first byte being the
char code, and not part of the bitmap data.