Tools - remove ps2ascii.ps, recode scripts to use txtwrite
ps2ascii.ps uses DELAYBIND, which we would like to get rid of, in
addition it doesn't do a terribly good job of actually extracting text.
The txtwrite device is more sophisticated, although it does not have the
added 'COMPLEX' feature of ps2ascii.ps which emits strokes and fills
as rectangles.

If the COMPLEX feature set should prove useful enough it would be easy
to add this functionality to a new device, but in reality I doubt
anyone is using this. The scripts specifically set -dSIMPLE.

This allows us to deprecate DELAYBIND and in future remove it.