pdfwrite - don't arbitrarily alter the 'current' shape alpha
Bug 697472 "Transparent box appears lighter"

Whenever we opened a substream, if it was an XObject resource we would
set the blend mode and shape alpha to the default values.

This looks wrong to me, and it means that when using a complex ExtGState
which includes a SMask, if there was an alpha other than 1 in the
ExtGState then we would write the constant alpha out twice, once
in the page content, and once again in the resulting Group Form
XObject which we create. If the alpha was not 1, then this causes it
to be applied twice, resulting in rendered content which is too pale.

Both of these look incorrect, but removing the blend mode causes other
files to fail. Removing the shape alpha causes progressions in a number
of test suite files.