Bug 697983: (re-redux) Drop zero length contours from glyph outlines
Zero length contours cause issues (line-caps) when the outline is stroked.

Turns out the same problem can arise with arcs as with straight lines.

That complicates things where TTFs are concerned since we have to convert cubic
to conic splines, and the floating point errors make it dubious checking for
point equality after the conversion. As a result, the checking has to be done
in the scaler specific interface code, rather than in the scaler independant
FAPI code.

That just means if we ever integrate another scaler that can handle
downloadable fonts, we need to take care to include the same checks as are
now in the Freetype interface code.