pdfwrite - fix /Colorants when converting DeviceN colour spaces
Bug 698129 "Produced PDF that gives "Wrong operand type" in Acrobat for some pages"

The supplied example uses a Duotone witha CMYK alternate and a Colorants
array (for the Pantone colour only). When converting to a different
space we preserve the Spot colours (and therefore DeviceN) but we
convert the alternate space to the specified base space by sampling the
original space.

However, if the DeviceN space contained a Colorants array we were writing
the new colour space array with the Colorants inserted in the wrong
position, immediately after the ink array instead of at the end.

This commit moves the Colorants entry to the correct place, and also
adds the (previously missing) Subtype key to the Colorants dictionary.