pdfwrite - prevent substitution of an SMask for an identical parent image
Bug #698128 "PDF created by gs generates "recursive /SMask" messages when run with gs"

When we have the (highly unusual) case where an image dictionary has an
SMask entry, and the SMask is an image which is identical to the parent
in every respect (dictionary contents, stream data) except for the
SMask key, it was possible for pdfwrite to decide to use the SMask
instead of the image.

This leads to a PDF which has an image which references an SMask in the
image dictionary, and the SMask is the same image XObject.

In this commit, when checking for an identical image, if the image we
find is the same as the 'current' softmask image, then do not treat it
as identical. This prevents us inadvertently detecting an image and its
SMask as being the same image.