Bug 698188: (clist) prevent reading off the end of a bitmap buffer
In clist_copy_mono, when the bitmap is clipped (in the x-direction) we apply
gross clipping to the bitmap data - the "clip" the left edge of the bitmap
to the nearest byte.

The problem is that we then calculate the aligned bitmap raster from that
byte offset. Thus, in the uncompressed case, if we copy the aligned raster
bytes for each scanline, we *can* end up reading off the end of the buffer.

Instead, use width/8 bytes as the amount to copy for each scanline. Then, if
necessary, zero the padding bytes.