PDF interpreter - detect and ignore recursive Form XObjects
Bug #698226 "Error: /rangecheck in --run--"

The PDF file supplied in the bug is broken, it has a Form XObject:

285 0 obj
<<
  /BBox [ 294.572 797.586 306.188 793.466 ]
  /Group 287 0 R
  /Matrix [ 1 0 0 1 0 0 ]
  /Resources <<
    /ExtGState <<
      /GS0 32 0 R
    >>
    /XObject <<
      /Fm0 285 0 R
    >>
  >>
  /Subtype /Form
  /Length 58
>>
stream
q
0 Tc 0 Tw 0 Ts 100 Tz 0 Tr []0 d
/GS0 gs
0 TL/Fm0 Do
Q
endstream
endobj

The file throws an error in Acrobat, Ghostscript recurses the Form until
it hits an execstackoverflow, and is then unable to proceed. Testing
Acrobat we see that it can detect recursion even when its not a simple
self-reference, a child referencing a parent is also detected and
ignored.

In the PDF interpreter we normally push a dictionary on the operand
stack and record objects as we encounter them, checking each new one
against all the ones executed so far, and exit if we find reuse. This
approach doesn't quite work for Form XObjects, because we cannot be
certain of the operand stack shape when a Form is executed, and so we
cannot locate our dictionary.

Instead we now store the dictionary in the top level 'pdfdict' which is
always available. The technique is otherwise the same.

I'm not totally happy with this, but its the best compromise between
performance hit, memory use and code frailty I could come up with.

No differences expected.