Update the nulldevice for DeviceN
Bug #696383 "Seg faults found by fuzzing in get_num_pdf14_spot_colors (gstrans.c:720)"

The problem here is really an example of our poorly maintained device
API. As stated previously the comments in the code seem to indicate
that all devices should either implement their own versions of all
the device methods, or use the graphics library default method. No
device method should be NULL, so it should never be necessary to check
this.

Sadly, many of our devices have NULL entries for some methods, and in
this case the nulldevice did not implement a method for
'ret_devn_params'. Presumably this was missed when the mthod was added.

In this commit, just up date the device to use the default method.