Bug 696387: iccprofile ref counting in gdevp14 compositor
The transparency compositor does extensive swapping and creating of ICC
profiles but was less than well disciplined with the reference counting. This
moves *most* of the profile swapping to a consistent use of the reference
counting.

The one remaining exception is the handling of colors for soft masks, which are
still handled "specially".