PDF interpreter - permit Widths arrays to contain indirect objects (!)
Bug 698275 "GPL Ghostscript 9.21: Failed to interpret TT Instructions"

The actual problem is that the PDF file contains a Widths array where
one or more of the entries is an indirect object. Since this has no
useful purpose beyond making the PDF file larger I can't think why this
has been done.

Nevertheless, as usual 'Acrobat accepts it' so now we do too.

The complain about the TT instructions is because in addition the file
contains two TrueType fonts which are broken.