pdfwrite - don't handle some kinds of x/yshow with TJ
Bug 698218 "Font matrix woes, also with PDF Reader/Writer"

In fact this is not a problem with the FontMatrix. The problem is
actually to do with the way the PDF interpreter executes certain kinds
of operation, basically Fonts where the /Widths array entries don't
match the actual glyph widths.

In this case we use an x/yshow operation with the modified width, and
this works perfectly well for rendering. However, for pdfwrite, we can
try to use the TJ operator, which works in a slightly different way...

xshow takes a displacement which is measured from the origin of the first
glyph to the origin of the following glyph, TJ by contrast adds a
displacement after the first glyph is rendered. Normally we essentially
subtract the width from the displacement to find the TJ value. However,
if the Widths array overrides the glyph width, and the actual glyph
width is significantly different (eg 0), then what happens is that
we apply the modified width after rendering the first glyph, and then
apply the TJ displacement. In effect applying it twice.

Detecting this condition turned out to be essentially impossible, there
are too many cases in PostScript files where similar trickery is pulled
to reliably determine that this occurs. It is, however, rare, especially
in PDF files, because very few fonts are neither horizontal nor vertical
in their FontMatrix (shearing is not uncommon but that's different)

So this commit adds a new boolean which we set false only when we have
a x/yshow operation, and the width and real width do not match (which
happens when we have a /Widths mismatch). This prevents the TJ being
applied and uses direct text positioning commands instead, which works.