Fix Bug 695333: PDF 1.4 transparency did not support Spot Color clist devices
Not just the tiffsep device named in the bug, but the psdcmyk device would
also fail if the clist was used. This file needs to be in the regression suite.

Fix by adding a pdf14_clist_copy_planes proc that forwards to the underlying
clist device. NB: The pdf14 rendering device already had a copy_planes device
which is used in page mode or during clist playback.