Bug 698034: Fix various problems with MFS downscaling to cmyk.
Firstly, the mfs buffer was not large enough.

Secondly, the downscaling logic in down_core4_mfs was broken - fixed
by copying it from down_core4.

Thirdly, the sense of the mfs was inverted (we had forced the white
spots to be zero).

All in all, my finest work. :(

Fixed here.