pdfwrite - initialise some new downscale parameters
Bug #698329 "Floating point exception in s_IScale_init()"

Commit a936cf76a33389610b558ffaf5de7389265fd657 to fix Bug #693684
"Add InterpolateControl parameter to limit image interpolation." added
a new parameter 'abs_interp_limit' but omittted to initialise the
variable in pdfwrite's use of the filter.

This led to an unuinitialised variable, which could eventually cause a
divide by zero fault.

It seems this parameter is only used when scaling up, so this commit
simply sets it to the minimum value of 1. At the same time initialise
a couple of other parameters too, these are probably unused and in any
event should be initialised by the filter code, but its best to be safe.