PDF interpreter - support for PDF 2.0 add HTO
I doubt we will ever see a Halftone Origin for real. According to the
specification its like a halftone phase but with a different co-ordinate
system.

For now simply treat its as a halftone phase, if anyone ever complains
we can address it, because they will be able to supply an example to
work from.