PDF interpreter - PDF 2.0 support add UseBlackPtComp
Michael already implemented the use of black point compensation, and
we have a PostScript operator to turn it on and off already (GS
defaults to on).

This just hooks up the PDF 2.0 graphics state entry to the existing
framework. Note that the PDF entry is a boolean whereas our existing
PostScript operator takes 0 or 1, so there's a little fiddling to get
that right.