PDF interpreter - PDF 2.0 support - per page OutputIntents
The only way previously to handle *any* OutputIntents array was to
specify UsePDFX3Profile and give the array index of the required
profile.

This is too inflexible as each page can have a different array, so the
indices may be different on each page. Add a new switch -dUseOutputIntent
which takes a string argument.

If we have set -dUseOutputIntent, then check teh document level and each
page for an array of OutputIntents. If found, check each OutputIntent
dictionary for a OutputConditionIdentifier and compare it with the
specified string. If the OutputConditionidentifier is 'Custom' then use
the value of the Info key instead.