Refactor gen_ordered to allow for use by an application (ghostscript).
Also Refactor htsc_save_tos to extract htsc_mask_to_tos to convert
final_mask.data to turn_on_sequence.

If GS_LIB_BUILD is #defined, the "main" and functions that write files
will not be included, and all but a few of the functions will be made
static. Exported functions are:
    htsc_set_default_params
    htsc_gen_ordered

Also ALLOC and FREE are used to allow the use of a memory allocator
(such as ghostscript's) and PRINTF and EPRINTF macros are used to
allow for redirection of messages.