Bug 698356: Add reference counting for SMask ICC profiles
Previously, we didn't properly reference count the ICC profiles for
soft masks because we simply swapped in the ones for the mask, and swapped out
the ones "in force" at the time into temporary storage, then swapped them back
when the SMask was complete.

That works fine *if* the PDF is well formed, and thus the SMask completes. But
it causes havoc if the PDF is invalid and causes an error during the processing
of the SMask.

We'll now properly reference count the profiles.