PDF interpreter - Cover some cases of illegal /Parent entries
Bug #698372

The PDF file has a Page with an invlalid /Parent (it references the Page
object). Attempting to use that Page object fails when tracking Resources.

This commit alters the object storage introduced with the PDF unique
font identification so that *any* dictionary with a /Type gets its
object number stored inside it. We can then compare the object and the
Parent when tracking resources and refuse to follow a self-referencing
object.

No differences expected