Fix FAPI memory leaks.
Noticed while chasing down other memory issues.

Two issues, as it turns out:

Firstly, descendants of a composite font were not having the notify method set, so
we can free the FAPI related memory for the font.

Secondly, the Freetype incremental interface object created for each font was not
freed along with the font object.