Coverity ID 158375
The function xps_set_nocache() is pointless, it doesn't do anything at
all, it merely retrieves a value from the PCL state stored in the
current instance. I suspect the value retrieved is garbage, since this
is XPS, not PCL, but since we never even use it, its irrelevant.

Whatever the original thinking behind this function was, its not doing
anything useful, so simply remove it.