Coverity IDs 158372 - 4, 158384
This is an attempt to work-around Coverity complaining about a 'tainted
 scalar' or 'tainted argument'. This is caused by reading a 16-bit value
from the Zip central directory and then using it without checking the
bounds. Since the legal bounds are 0 - 2^16 its rather hard to see how
we can bounds check this....

If this doesn't silence the warnings tomorrow, or creates new ones
complaining that the bounds check is pointless (this has happened with
Coverity before) then I'll undo this commit and just tell Coverity
to ignore the warning.