Coverity IDs 158378, 158393
Add some limit checks to see if they silence Coverity. Also revisit the
check added in commit c5bf24912f589c881be44408e110ebdbb2fcd5b9 to
correct the limit check.

Coverity complains about reading values from a file format and then
using them without validating them. Its hard to see how to validate
values which can legitimately range over the whole space available. If
we add checks for that, then Coverity complains that the checks are
pointless (which is true of course).

In these cases, however, we are reading unsigned shorts into signed
integers, so we can add a (pointless, really) test to ensure that the
integer value lies in the range of a short. This should satisfy the
'tainted scalar' complaints and also should silence the overflow
warnings when using the values to allocate memory.

Normally I wouldn't do this as is costs performance, but we only execute
this code once when we open the file, so its negligible.