pdfwrite - don't attempt to use the image-to-mask device unless its present
Bug #698390 "gswin32/64 crashes during PDF conversion"

This looks like a simple oversight, other places where we attempt to
use methods from the mask device we first check to see if we have a
mask device present, which we do not in this case. This commit simply
checks the presence of the mask device and doesn't attempt to use it
if we have not installed one.

No differences expected