pdfwrite - add code to apply transfer functions instead of preserving
With the release of the PDF 2.0 specification transfer functions have
been deprecated in graphics states. This means we can no longer simply
preserve transfer functions as we have done in the past, and removing
them can lead to unexpected results (Quark XPress is notorious for
using transfer functions inadvisably). So this commit alters the colour
handling code fairly extensively in order to allow transfer functions
to be applied.

In order to achieve this there are conditions under which we are forced
to convert colour spaces, as its not possible to apply the transfer
function reasonably to the base space (CIEBased colour spaces are an
example of this). In this case we will use the ProcessColorModel in
effect at the time as the space to convert to.

Its also (unfortunately) the case that when the input is a PDF file,
the PDF file uses transparency and the PDF file has transfer functions
set in the graphics state, that some kinds of operation (soft mask
groups for example) cannot have the transfer function simply applied.
This can result in surprising output. In these (admittedly rare) cases
removing the transfer function causes differently incorrect output.
There is no realistic way for us to take a PDF file like this and
produce an equivalent PDF file which does not contain transfer function.
Worse, we can't even detect this condition in advance.

This concludes the work required to support PDF 2.0 creation, and
Ghostscript should now be able to create and consume PDF 2.0 files
though additional work may be required, and we may choose to add
more functionality (eg Document Parts processing) in the future.