PDF Interpreter - Paint one pixel for Patterns with BBox width/height of 0
Bug #698431 "PDF 2.0 - Patterns with BBox where width or height = 0 should paint 1 pixel"

The ISO32000-2:2017 specification explicitly states (See section 8.7.3.1
Table 74 and section 8.7.4.3 Table 77) that a Pattern with a BBox where
the width or height is 0 should still paint one pixel.

We achieve this here by detecting the condition where width or height in
a Pattern BBox is exactly 0, and add a tiny amount to the urx or ury
to ensure that one pixels is touched, and therefore painted.