PDF interpreter - fix implementation of HTO
I implemented HTO by using the existing sethalftonephase without an
example to look at. It turns out that sethalftonephase uses
.setscreenphase which expects three integers while HTO consists of
an array of 2 reals.

We already manufacture the third integer (the texture) and use aload to
extract the two reals. This commit uses transform to apply the CTM and
then cvi to convert the HTO numbers to something that makes a difference
and is legal.

Fixes the error and the resulting file (rendered at 1bpp) does show a
difference in the halftone at 300 dpi or better.