clang warnings in the TTF code.
There was a type mismatch: we were using a boolean to track if an error had
occurred, but in most places we were treating it as a gs error code (int type).

Change the type to int, and tweak the couple of places where it was treated as
a bool.