PDF interpreter - add support for Choice fields in Widget annotations
Bug 698461 "PDF files with filled forms are not rendered correctly if they have drop-down lists"

We didn't support Choice fields in either AcroForms or Widget annotations
which leads to these not being rendered if they have no Appearance (in
the case of widget annotations) or not rendered at all in the case of
AcroForms.

This commit adds simple support for single selection Choice fields, we
may want to improve this later.