Coverity ID 94708
The original code in here was correctly noted by Coverity as being
dead, because the division by BJC_RESOLUTION_BASE meant that we were
always comparing 0 with 0.

In addition, the use of bit shifting meant that the resolution wasn't
tested against a simple multiple of 90 dpi, but against powers of 2
times 90 dpi, which rapidly led to ridiculous values.

I'm not certain what the original limits should have been here, and
I don't have a bjc device to test against, but the code as it stood was
clearly incorrect, and this at least stands some chance or working.