Coverity ID 94943
I think this code is incorrect, because this routine *always* maps
CMYK to CMYK, so there's no need to test the colour depth.

Also, it always gets provided with a CMYK buffer, not an RGB one so it
actually can't write off the end of the buffer. But the declaration
was wrong, so fix that here to avoid the warning.