Coverity ID 94814 and 94980
I wasn't able to find a way to reproduce this, because it relies on
a particular sequence of lines/points. We must fill the data buffer to
the point where another line or curve will not fit, and then the next
curve or line must be a 'POINTS_NONE'.

In this case we will not flush the accumulated points, which means we
will not reset the counter, but we will then still go on to save the
current co-ordinates in the data buffer, which will cause an overflow.

Fixed by reducing the 'full' check slightly, so that we still have one
line/curve slot open in the data buffer. If we get a POINTS_NONE then
we check to see if we have already gone past the slop. If we have, simply
reduce the counter by one which will cause us to overwrite the previous
points. Since this must have also been a POINTS_NONE, this just means
in effect that we have combined consecutive ones.

I'm not certain its even possible to get consecutive POINTS_NONE, but
its best to be safe.

This should silence Coverity, and fixes the genuine (if hard to
reproduce) potential problem.