Coverity IDs 107340, 107341, 107343, 107344, 107345, 121455
Considerable rework of error checking in this device.

Firstly add error return checks on fseek and fwrite. Alter functions to
return those errors. Alter the parent functions to check the errors,
repeat until we bubble up to the top level.

Coverity claims that ftell() can return a negative value, which as far
as I can see is not true. However, by casting the return value to an
unsigned type, and altering the variable types appropriately, I think
we should eliminate this warning.

Refactor conditionally compiled code to eliminate a variable and get
rid of a dead code warning.