Coverity ID 94489
The spec says that it must be 3 bytes, and uses a 16 bit shift example
(suggesting it is stored in the Brother driver as 24 bits).

For us its a short, so rather than copy the brother code, set it to 0 to
avoid any confusion over whether shifting a short right by 16 results in
all 0 or all 1.