Coverity ID 158391
Initialise 'backtable'. The code does later fill in some entries in
backtable, but it doesn't fill all of them and I can see no coherence
between the filling in code and the code that reads data from it. It
really does look like its possible to read uninitialised data here.