Coverity IDs 94596, 94726
A debug print was identical in both the if and else clauses. Not really
important since its debug, but change the else to match others of the
same type.

line_ptrs was calculated and altered in the loop invocation, but its
then immediately overwritten in the loop body, making it pointless.