Coverity IDs 94719, 94889, 94912, 94967, 95088, 101187
Return an error instead of storing it (where it gets lost)

As per gdevpsd.c and gdevprf.c refactor the #if conditional compiled
code to remove a variable and get rid of Coverity warning.

In the 'build_cmyk_raster_line_fromplanar* routines, pass the params as
a pointer instead of copying the whole 500+ byte structure around.

Finally, not a Coverity warning but a compiler warning; remove an unused
variable.