Coverity ID 94479
The macro used here says it shouldn't be used for run-time calculations.
However this code does use it, I'm not certain for what purpose. Since
it clearly is possible to get a negative shift, add a guard to the
values to make sure that can't happen.

Since this code is conditionally compiled out, anyone who compiled it
in and finds a problem can easily fix it.