Coverity IDs 94478 and 94485
The first is a clear typo, the code below this point is the same as the
code modified in this commit.

The second one is less clear, but I think its reasonable, if there is
no '0' (-1 return from next_zero()) then we can't make the feature
larger by darkening to the right, so darken to the left.